example test cleanup: remove unneeded thread include from client
cpplint bans the use of this header. It's not actually needed in
client.cpp, so this changes removes it.

It is left in place in replica.cpp, because it is needed there. The
reason cpplint bans the use of `#include<thread>` is that cpplint is
specifically written for the Chromium project, which implements its
own thread header. This may mean that we want to go ahead and fork a
copy of cpplint that we can then modify to fit our specific needs
(since we're not using Chromium's headers).