Add key generation utility for Concord
This commit adds a key generation utility for concord; it can be called
from the command line to generate all RSA and threshold signature keys
needed by a Concord deployment, which it outputs to keyfiles. A function
for reading the keyfiles back in is also included, as is a test of the
key generation utility and a README documenting how to use it.
Furthermore, this commit makes revisions to simpleTest so that it uses
keyfiles of the new format. The key generation utility is implemented in
the tools subdirectory. See tools/README.md for more information about
this utility.