Add code that retransmits StartSlowCommitMsg message.
Problem: If several StartSlowCommitMsg messages are lost, we won’t start the slow path, and we will have to wait until the view change.
Solution: add a flag to ReqMissingDataMsg (“true” for sequence number k IFF the replica has started slow path for k). Use this information to retransmit StartSlowCommitMsg.