Add Googletest as a dependency in a submodule
I added gtest as a submodule so we can start to write unit tests.
CMakeLists.txt was updated to pull it in to the build. A very basic
example test was added in bftengine/tests/datastore. The README was also
updated to reflect that submodules need to be initialized.

Additionally, vim temporary files are now ignored by adding them .gitignore.