Clang-format: Add support code for auto code formatting
This is a port of work originally done by @robem for our closed source
stuff.

Add clang-format config file

The config file introduced in this change will be the base for all future
clang-format runs. It was generated with the following command.

  clang-format -style=Google -dump-config > .clang-format

Three modifications were made to the generated config. We don't sort
includes, and we don't binpack parameters or arguments.

Add a wrapper script for running clang-format

We are interested in C/C++ related files only and hence we need to filter all
files under concord-bft. This change introduces a script that runs
clang-format with the proper arguments for you and is written with CI/CD in
mind. If you provide `--is-required` to the script then it will return 0 if no
code format change is needed and 1 otherwise.

A cmake script was added for the script, but not integrated into the
top-level build. We can integrate it later when we have a more flexible
top-level build with multiple targets.

Also add some documentation about installing clang-format.