Enable setting build options from the cmake cli
Rather than forcing users to edit CMakeLists.txt everytime they want to
change a build option, we enable key options to be set from the command
line using cache variables.

Note that for string variables we force a default only if the string is
unset or set to a falsy value which doesn't make sense for non-boolean
options.