Fix local build when missing deps/gtest
A prior fix was made to only download the deps/gtest submodule if it
wasn't already downloaded to fix a project using concord-bft that
already had initialized submodules. Unfortunately, the glob pattern used
to perform the check for an existing dep was incorrect and only checked
to see if the directory existed and not if there were any contents. This
commit fixes the glob pattern.

Additionally, unprivileged users on github cannot use git paths to pull,
even from opensource repositories. Therefore replace the .gitmodules
path previously starting with `git@` with an `https://` style path.

I tested this locally and inside our other product build to ensure
correctness. I also tested it on a new server without github
credentials.