Add a metrics library (#69)
The metrics library provides common data types with well defined
interfaces to allow us collect information about the running bft node.
There are two container types, and three value types included in the
library.

Components
---

A `Component` represents a specific piece of code that needs its metrics
trackable during runtime, such as the ReplicaImp or BcStateTransfer. A component
contains a fixed set of Values (Gauge, Status, Counter), is updated on a single
thread, and is optimized for fast writes. The goal was to make writes as fast as
possible while still allowing for some semblance of generic behavior, so that
stats from different components can be aggregated and reported on in similar
ways. They are write optimized, because updating can happen very frequently
especially in the case of counters and histograms (coming in a future commit),
and we don't want to limit our collection due to overhead.

We make writes fast by returning Handles to values which wrap indexes
into vecs. Therefore an update is essentially a vec indexing followed by
the actual operation. Handles are expected to be kept as member
variables in the bft type that also holds the component. Therefore, stats
updates occur by updating member variables, as would be done in a more
naive system.

Aggregators
---

There is typically one aggregator per node which provides an eventually
consistent view of the metrics of all components and their contained
values. Aggregators are expected to be used to retrieve the metrics of
the entire system during end to end tests, and to optionally provide
live runtime debugging information for a given node. They should be
accessed much less frequently than components.

Aggregrators are updated periodically with component values in a thread
safe manner via `UpdateValues`. We minimize the amount of data transferred
during update by only transferring values and not including their names. Since
all components have a pre-defined number of values when they start, we can
inform the aggregator about the names when we call `RegisterComponent()`. This
is the reason that value names are separated from the value contents into
separate types (Names and Values) containing vecs for each value type. We could
always further minimize the amount of data transferred by keeping track of which
values were updated and only transferring those, but that would likely make
updating take longer, and adds complexity. We can do this later if measuring
shows performance issues.

Values
----

There are 3 value types (stats types) that define the types of metrics being
collected. A `Gauge` reports a static integer that changes discretely, but not in
an incremental fashion. A `Status` reports any short textual data, such as whether the
replica is currently a primary or backup. A `Counter` is used for incrementing
stats, such as the number of sent messages. Histograms (via HdrHistogram) are
expected to be added in a future commit.

---

The included `metric_test` tests the basic functionality of all the types and
shows how to use them. Tests can be run with `ctest -R metric
--verbose` from the build directory.

Future commits will include serialization, as well as integration of
components and aggregrators into the rest of the codebase. After that we will
need to add a client command to retrieve the serialized metrics so we can use it
for testing and runtime inspection.