Track aggregation timestamp in component values
This will allow us to display the time when the metrics are from for
each component once we add serialization and a server.

Note that we use the system clock here because we are going to want a
human readable time. It's possible that this isn't monotonic, and so the
test can possibly fail. That'd be pretty damn unlucky though. If we
think it's a problem I can just remove the testing, as it's not strictly
necessary.