Metrics: Add JSON serialization support
A simple hand written JSON serializer was implemented to serialize an
Aggregator and its corresponding components and metrics. There is no
deserializer as deserialization is expected to be done elsewhere as part
of the metric server clients. Since serialization is so simple here, we
decided we didn't need to pull in any dependencies and could hand roll
it.

Tests can be run via:
ctest -R metric --verbose

The JSON output generated by the test as pretty printed through json_pp
looks like this:

     {
        "Components" : [
           {
              "Gauges" : {
                 "total_peers" : 4,
                 "connected_peers" : 3
              },
              "Statuses" : {
                 "state" : "primary",
                 "commit_path" : "SLOW"
              },
              "Counters" : {
                 "messages_received" : 1,
                 "messages_sent" : 0
              },
              "Name" : "replica"
           },
           {
              "Gauges" : {
                 "blocks-remaining" : 4
              },
              "Statuses" : {
                 "state" : "sending-blocks"
              },
              "Name" : "state-transfer",
              "Counters" : {}
           }
        ]
     }