pyclient: Use hybrid req_seq_num based on clock
Sequence numbers for the python client now consist of a wallclock
component (time since epoch in milliseconds) and a 22 bit counter. This
allows clients and the machines they run on to stop and restart and pick
up with a greater sequence number.

Note that since this sequence number is based on wall clock time, it is
possible after a client restart that it will not be monotonic. In this
case the request will be rejected by SBFT as stale. However, it will
quickly fix itself as the clock adjustments should not be very large.
This is unlikely to be an issue in practice.

While we could use the monotonic clock to prevent the above problem due
to restarts of the client, we use the wall clock to prevent the problems
of the machine restarting. In the case of a machine restart the
monotonic clock gets set back to 0 and is useless for our needs.