Bugfix: Fix overflow of counter in SimpleClient
When the SimpleClient counter in a request sequence number hits its
maximum value of 0x3FFFFF, the time part of the sequence number
increases by 1 millisecond. However, the counter part was not getting
reset to 0, but instead was incremented also. If this scenario occurred the
an assert would fire and exit the program. The code was fixed to match
the invariant in the assert.