Collect metrics inside ReplicaImp (#75)
* Collect metrics inside ReplicaImp

ReplicaImp now contains a metrics Component and registers some metrics.
The metrics get updated in both ReplicImp.cpp and
ControllerWithSimpleHistory.cpp.

Because the Component requires an aggregator, but an aggregator is not
yet instantiated for the bftengine as a whole, we create a dummy
aggregator, but never actually update it. To avoid having to change the
Replica constructor we also create a method that allows setting a
different aggregator and forward that onto the component in the Metrics
code so that it can register the component with the new Aggregator.

Since the metrics themselves are references into the component they must
be class members that are part of initialization. They will therefore
always be updated and part of ReplicaImp. The plan is to make the
feature optional by not always allowing the aggregated data to be
retrieved. In other words the to be built metric server will not always
be enabled. The overhead of tracking these metrics is very low, so this
seems like a reasonable tradeoff. Otherwise, the implementation of the
Metrics library itself may need to be changed.

A future commit will contain an optional UDP metric server which also contains the
centralized instantiated Aggregator. It will also contain code that
calls metricsComponent_.UpdateAggregator() periodically on some timer
tick.

* Don't update metrics in controllers

Controllers now return when the currentFirstPath changes so that the
metric can be updated in the replicas.

Also, each external received message now has it's own counter metric and
there is a separate counter for internal messages.