Add a python metrics client for testing
A simple UDP python client was created for testing. It uses trio so that
we can run concurrent requests during testing and it interroperates with
the bft client.

Additionally, client types were moved into config.py and test_client was
modified to reset its working directory so as not to break tests that
run after it.