Start the UDP MetricsServer for simpleKVBC tests
The Replica abstract base class (along with ReplicaInternal and
ReplicaImp) now has an additional method, `SetAggregator` that allows
establishing the metrics aggregator to use for a given replica. This
aggregator is a member of the MetricsServer class which runs in main for
each SimpleKVBC replica. To allow configuring the SimpleKVBC replica,
the aggregator was added as a parameter to the constructor.

As part of this change the CMakeLists.txt for simpleKVBC was cleaned up
so that it no longer has to include a bunch of relative include
directories for bftEngine, threshsign, and logging. It now just uses
`target_link_libraries` and gets them all for free. This is the modern
cmake approach.

Additionally, a `using namespace SimpleKVBC` directive was removed from
simpleKVBCTests.h as using directives should not be in header files.
This using directive was moved to the cpp file instead.