SimpleKVBCTests bugfix and cleanup
Graceful shutdown is not currently supported. When an assert fires and
the program terminates another system error occurs during shutdown in
the signal handler. This causes an endless stream of `terminate called
recursively` messages to display to the console. As the signal handler
is unnecessary for this test, remove it.

Additionally, change the CHECK method into a macro so that we can see
which line the assert occurs on.

Also fix a typo: `Verion` -> `Version`.