Add automated system test framework in python
Add an automated system test framework in the test directory, along with
the first automated system test, `test_state_transfer`. Please see
`test/README.md` for more details about the design of the framework and
how to use it.

This test infrastructure relies on retrieving metrics from replicas
running concord-bft with an application specific state machine. This
commit also ensures that the aggregator in the metric server gets
updated periodically by adding a new timer to ReplicaImp. This ensures
the tests get valid metrics when using bft_metrics_client.

Additionally, the number of client proxies was increased in the
TesterReplica so that we could use more concurrent clients without
requiring a configuration file. A bug was also fixed such that the
configured number was used instead of the replicaParams version.

`statusReportTimerMillisec` was also shortened significantly in the
SimpleKVBCTests code so that nodes can discover each other more quickly
after a failure or partition.