Fix incorrect connection attempt timeout for TLS module (#85)
* Fix incorrect connection attempt timeout for TLS module

This commit aresses incorrect behaviour of the connection timeout management handling in the TLS module. When the outgoing connection is created and async_connect is started,
the deadline timer is started in parallel and has race condition with the connect_complete callback. We noticed on the slow machines that there may be a case when time ticks before the connection is opened, but connection is already open when we try to reset the socket for another attempt.

The fix is to start the timer after failed connection attempt and to delay the next connection attempt. Moreover, there were few places in code where the handlers where binded to 'this' raw pointer instead of using 'shared_from_this()'. This is fixed as well.

* fix code review comments and simpleTest.py bug

* stop async_accept on operation_abort error