Add a sequence number to MetricsServer (#87)
We want to be able to ignore stale replies during tests, so change the
protocol used by the MetricServer to include a sequence number.

Also change the bft_metrics_client to only operate against a single
replica. This allows easier tracking of sequence numbers and ensures we
don't have to buffer replies in any way.

All tests were updated to reflect the new code. Tests pass.

    andrewstone@ubuntu:~/concord-bft/build$ ctest -R '(pyclient|metric)'
    Test project /home/andrewstone/concord-bft/build
        Start  9: pyclient_tests
    1/3 Test  #9: pyclient_tests ...................   Passed    1.46 sec
        Start 10: metric_tests
    2/3 Test #10: metric_tests .....................   Passed    0.01 sec
        Start 11: metric_server_tests
    3/3 Test #11: metric_server_tests ..............   Passed    0.18 sec

    100% tests passed, 0 tests failed out of 3

    Total Test time (real) =   1.66 sec