Fix error where metrics component was unregistered
In some binaries, a metrics server will not be run, and therefore there
will be no in use aggregator set via a call to
`ReplicaImp::SetAggregator()`. However, we still update metrics with a
default aggregator and the update timer still fires in ReplicaImp. If
the the replica component doesn't register itself with this default
aggregator, an `out_of_range` exception will throw from a call to
`components._at(component_name)` when the timer fires.

This was occurring periodically during test runs of
pyclient/client_test, because the simpleTest replicas don't run a
metricsServer. This commit fixes the issue.