Add support for testing concurrent operations
A new component to the python testing framework, SkvbcTracker, has been
added that is capable of tracking requests and replies for messages sent
to a cluster of SimpleKVBCTests TesterReplica nodes. Tests can use this
component to record requests and replies and then verify that all
returned replies can be placed in a total order that makes sense. In
other words all the operations are linearizable. The tool itself makes
no promises about finding all possible violations, although it will not
provide any false positives.

The SkvbcTracker lives in skvbc_linearizability.py, and significant
documentation for it lives in the doc comments, particularly the doc
comment for the SkvbcTracker class. Reviewers should read and
understand that comment prior to proceeding with review. Tests in
test_skvbc_linearizability show how to use the tracker and should
provide more clarity.

Additionally, many specific exceptions were created to indicate
linearizability violations. These live in skvbc_exceptions.py. Other
exceptions for skvbc testing may be added there in the future.