Add -s CLI param to SimpleKVBC TesterReplica
Add a new -s parameter that makes statusReportTimerMillisec
configurable.

We should probably have an automated mechanism for writing and reading
bftengine ReplicaConfig structs to/from files in standardized formats
and removing all other duplicate configuration structs. After talking
with Igor, we decided that simply adding a parameter was the right thing
to do at this moment in this PR.