Add automated system test framework in python (#89)
* Add automated system test framework in python

Add an automated system test framework in the test directory, along with
the first automated system test, `test_state_transfer`. Please see
`test/README.md` for more details about the design of the framework and
how to use it.

This test infrastructure relies on retrieving metrics from replicas
running concord-bft with an application specific state machine. This
commit also ensures that the aggregator in the metric server gets
updated periodically by adding a new timer to ReplicaImp. This ensures
the tests get valid metrics when using bft_metrics_client.

Additionally, the number of client proxies was increased in the
TesterReplica so that we could use more concurrent clients without
requiring a configuration file. A bug was also fixed such that the
configured number was used instead of the replicaParams version.

`statusReportTimerMillisec` was also shortened significantly in the
SimpleKVBCTests code so that nodes can discover each other more quickly
after a failure or partition.

* Fix error where metrics component was unregistered

In some binaries, a metrics server will not be run, and therefore there
will be no in use aggregator set via a call to
`ReplicaImp::SetAggregator()`. However, we still update metrics with a
default aggregator and the update timer still fires in ReplicaImp. If
the the replica component doesn't register itself with this default
aggregator, an `out_of_range` exception will throw from a call to
`components._at(component_name)` when the timer fires.

This was occurring periodically during test runs of
pyclient/client_test, because the simpleTest replicas don't run a
metricsServer. This commit fixes the issue.

* Remove signalHandler in simpleTest

Graceful shutdown does not work correctly yet. Due to any other
uncaught exception throwing and triggering a SIGABRT, this signalHandler
will throw over and over resulting in the message "terminate called
recursively". This hinders our debugging ability by making it hard to
see the actual output during the error. Scrolling for several pages to
see the error is required.

* Add -s CLI param to SimpleKVBC TesterReplica

Add a new -s parameter that makes statusReportTimerMillisec
configurable.

We should probably have an automated mechanism for writing and reading
bftengine ReplicaConfig structs to/from files in standardized formats
and removing all other duplicate configuration structs. After talking
with Igor, we decided that simply adding a parameter was the right thing
to do at this moment in this PR.