Add a concurrent test against a healthy cluster
Run a test with multiple configurations of a cluster and concurrent
clients sending requests. No faults are introduced in this test and it
is a baseline to ensure that without faults or view changes the cluster
returns valid data.

Also, rename the test group in cmake to skvbc_tests, and add this test.