Fix tracking of blocks for concurrency tests
When a CausalState is constructed as part of a request, we need to
snapshot the actual number of blocks missing between the last
consecutive block and last known block, rather than assuming all
intermediate blocks were missing during verification. This calculation
allows us to check relevant concurrently created blocks for
correctness when checking to see if a read of a given value should have
been possible. Without it, we may overestimate the amount of missing
blocks and check less blocks as possible sources of any values read.
This could cause false positives where reads are shown not to be
linearizable, even though they actually are.