Use unique_ptrs in InternalMsgsStorage
InternalMsgsStorage is used for transferring internal and external
messages across threads. This makes the queues use unique_ptrs instead
of raw pointers so that we don't have to keep track of ownership and
lifetimes cross thread. All callers and users were changed to reflect
the new types.

Note that this change causes no changes in the safety of the existing
code. It is being used an introduction to using unique_ptrs to indicate
transfer of ownership across function and data structure boundaries.
Future work will remove the need for manual deletes as the use of smart
pointers is continued.

This was stripped out of a much larger change that attempted to convert
all messages to being contained in unique_ptrs. That change was huge and
effectively unreviewable, so a piecemeal attempt is being made instead.