Format IncomingMsgsStorage.(cpp|hpp)
This is in preparation for another commit which makes this file take
unique_ptrs instead of raw pointers. I formatted the code to make the
changes easier, and this will allow an easier diff for the actual coming
changes.

The coming commit is the first step in using smart pointers in the
replica code path.