Add initial version of design document.
This document is intended to cover the overall structure and
implementation of the code as it currently exists. This initial commit
covers just the public interfaces to the bftengine. Future commits will
cover more details such as state transfer, bftengine / protocol
implementation, timeouts, persistence, etc...

It's possible that we'll want to turn this from a single document into a
developer's guide with different "chapters". For now, let's keep it
simple though.