use _expectedDestId instead of _destId to reconnect
The call to `_fOnError` in `dispose_connection` is what triggers
reconnection on failure, by calling `on_async_connection_error`. That
function uses the passed parameter to determine which node to
reconnect to.

If the most recent connection attempt did not make it through the
authentication phase, `_destId` will still be set to
`AsyncTlsConnection::UNKNOWN_NODE_ID`. (See `check_certificate`.)

This change passes `_expectedDestId` to `_fOnError` instead. That
value is set in the constructor of AsyncTlsTcpConnection.