ReplicaConfigSingleton
In order to access the replica configuration values from anywhere inside
concord-bft we can either pass around the ReplicaConfig structure, introduce a
global variable or create a singleton. The former isn't easy to do because it
would touch a lot of function signatures and hence would cause a lot of merge
conflicts for other developers. A singleton is pretty much a global variable
but with the guarantee of a consitent view. At the moment, the properties are
read-only but this is no blocker for future setters.
The main motivation is to make concord-bft properties configurable at start-up
time. Re-compiling every time we want to change a constant is simply not
practical. With the proposed approach, you can add a new config value to the
ReplicaConfig structure, add a getter in the singleton and use it wherever its
needed without passing around the value.
In a follow-up commit you can see its usage.