enable -Werror for compilation (#129)
* enable -Werror for compilation

While the build is clean, let's turn on warnings-as-errors, to help
keep it that way.

* Include Logger.hpp, not Logging.hpp

Logger.hpp decides whether to include Logging.hpp or
Logging4cplus.hpp, basic on the USE_LOG4CPP macro definition.

* make BlsPublicParameters non-global in TestThresholdSerialization

The static initialization of this global happens before the logging
subsystem is initialized, which causes a segfault. Moving this to
RelicMain moves the parameter init after the logging init.