Fix skvbc tests
Runnng `make test` was causing the skvbc_tests to fail. Running in
verbose mode, an assert was being triggered because persistence wasn't
setup properly.

```
15: skvbc_replica: /home/andrewstone/concord-bft/bftengine/src/bftengine/BFTEngine.cpp:123: static bftEngine::Replica *bftEngine::Replica::createNewReplica(bftEngine::ReplicaConfig *, bftEngine::RequestsHandler *, bftEngine::IStateTransfer *, bftEngine::ICommunication *, bftEngine::MetadataStorage *): Assertion `false' failed.
```

PR #130 was merged in without running the automated tests which caused
this failure. The fix was to use the FileStorage class as backing for
MetadataStorage when calling the SimpleKVBC::ReplicaImp constructor. We
also could have turned on debugPersistence and avoided the assertion,
but it's better to be running our tests with actual persistence. This
way we can start testing the persistence itself from the testing
framework.