Add a Timers class to the util library
`Timers` maintains a vector of timers, where each is associated with a
callback.  The `Timers::evaluate` method finds all expired timers and
runs their corresponding callback.

Timers can be either `ONESHOT` or `RECURRING`. A ONESHOT timer will be
removed from the collection after it fires once. A RECURRING timer will be
rescheduled to expire at the current time + the duration passed to it's
constructor. Durations can be modified by calling the `Timers::reset`
method with the handle returned from calling `Timers::add`.

The Timer and Timers classes use the std::chrono::steady_clock, and
utilize std::chrono types for safety, to reduce the occurrence of issues
like those in #141.

A detailed test showing how to use the API is also part of this commit.

Future commits will use the Timers class instead of the
SimpleTimerScheduler to provide a typesafe interface for callbacks,
and allow the removal of timer methods from the InternalReplicaAPI. We
can also remove the *TimerHandlerFunc methods from ReplicaImp, since
recurring timers will be rescheduled automatically.