Add rocksdb related storage code (#167)
All rocksdb code serves as an example and is completely optional.

The storage code is spit into two layers in separate directories:

 * `rocksdb` - contains rocksdb client code and MetadataStorage
 implementation
 * `blockchain` contains higher level code that uses the database
 interfaces to organize blocks of data such that they can be stored
 persistently.

The rocksdb layer is the absolute minimum storage code and can be used
independently of the blockchain layer. The blockchain layer is a thin
abstraction on top of the database interfaces that creates a useful API
for accessing persistent blockchain information. This code serves as an
example and is also useful for building a test replica similar to the
simpleKVBCTests TesterReplica that will allow us to create test clusters
full of persistent replicas.

The majority of this code was not written by me. It was ported from a
closed source repository. I only made minor tweaks to remove circular
dependencies, change namespaces, and split the code into multiple
libraries. Credit for most of the code goes to @guyg8 @robem @beerriot
@yuliasherman. Apologies if I forgot anyone.

Some types were extracted into our util library so that they could be
shared by both the blockchain and rocksdb libraries. It's anticipated
they will be useful elsewhere.

This change also includes tests for storage code and sliver that can be
run in the standard manner with `make test` and `ctest -R <TEST>`.

Please see the README in the storage directory for more information.