Add an in-memory db client to storage
An in-memory client was added as a new low level storage library. It
implements the db client interfaces as an alternative to the rocksdb
implementation. This allows the blockchain code to be built on top of
the in-memory implementation if rocksdb is not being used.

The cmake files were changed to decouple the rocksdb build from the
blockchain build. The README was also updated to reflect these and the
prior changes.