Use storage instead of test code in simpleKVBC
This PR is a deduplication effort for in memory storage code used by the
simpleKVBC test library. The test code consists of old copies of the
code in the storage library. The storage library contains the canonical
versions and is useful outside tests as well.

Of note is that we have replaced the Slice datatype, which aliases
raw pointers, with the Sliver data type that takes ownership of the data
in a shared_ptr. This requires copies in some places where they did not
exist before. However, this change is an intermediate step towards a
full refactor of SimpleKVBC, which is in progress. That will remove some
of these extra copies, as the signatures of some functions that take
slivers have changed to taking the raw pointers.

In order to make copying of buffers more ergonomic, a static `copy`
method has been added to the Sliver class, along with a test.

The code changes have been validated via existing simpleKVBC tests. The
scripts for manual testing have been run, along with the python testing
framework that uses the SimpleKVBCTests TesterReplica. All tests pass.