Reduce log level for select log messages.
This commit demotes a number of log statements in Concord-BFT from INFO to DEBUG log level. This log demotion is helpful because Concord-BFT has a number of log statements, at the INFO level before this commit, in common, frequently executed, and/or non-error code paths that do not provide much information for a normally operating system but which do clutter up the logs and increase their growth rate. The log messages demoted by this commit were deemed not to provide enough new information in the common case to justify the INFO log level. Additionally, this commit replaces the DEBUG_STATISTICS compiler macro (which controls whether Concord-BFT will automatically periodically output debug statistics) with a new debugStatisticsEnabled configuration parameter in the bftengine::ReplicaConfig struct.