Add an in-memory db client to storage library
An in-memory client was added as a new low level storage library. It
implements the db client interfaces as an alternative to the rocksdb
implementation. This allows the blockchain code and db_metdata_storageto
be built on top of the in-memory implementation if rocksdb is not being
used.

The README was also updated to reflect these changes.