Refactor SimpleKVBCTests and SimpleKVBC
Significant refactoring of the SimpleKVBC code and tests was done in a
private repository. This PR ports over the majority of those changes,
modifying headers and types where appropriate.

For tests, we serialize block metadata the same way we do with the rest
of our serialization code, by writing the types directly as a char* buf.

Additionally, this commit:
    * Adds a log message indicating whether storage is New
    * ReplicaConfig duplication was removed from tests. Tests now use only
      the bftEngine::ReplicaConfig.
    * Clearer warning logs were added to BcStateTran.cpp
    * Removed debug logs from blockchain_db_adapter.cpp, since they log
      every operation and cause tests to timeout with debug logging turned on.
    * Fixed the ordering of request types in skvbc.py since they were
      reorded in the modified message code ported over.
    * Reordered the message struct fields in simpleKVBTestsBuilder to
      match the originals, to prevent having to change the order in the
      python test definitions.
    * Added a `!=` operator to Sliver.

All tests pass using both the MemoryDB and RocksDB storage backends. The
tests are currently hardcoded to use the MemoryDB backends. A future
commit will create two separate binaries so that we can run all relevant
tests for both in memory and persistent codebases.