Fix reversed exit code in bftengine simpleTest client
SimpleTestClient::run() returns true if successful.  (It presumably could
return false if it failed, but it actually aborts the process.)  The caller
in client.cpp used this value as a process exit status, but the convention
for process exit status is reversed, where true (1) means failure and
false (0) means success.  This commit fixes the problem.

Fixes https://github.com/vmware/concord-bft/issues/185.