Include batching data in periodically logged DebugStatistics.
When debug statistics are enabled, this patch adds three more statistics to
those logged: the number of prePrepare messages sent, the average number of
requests in each of those prePrepares, and the average number of requests
pending that could not be included in each prePrepare.

Fixes https://github.com/vmware/concord-bft/issues/182.