Revert #195
The code is still incorrect below here, but it works correctly on 4 node
clusters. The change made in 195 broke several unit tests and also has a
critical problem of attempting to read from storage while a write
transaction is in progress, which violates invariants in asserts. In
fact the change in #195 led to a situation where there were 2 mutually
exclusive asserts in the code path, and thus would always abort on any
view change attempt:

```
bool PersistentStorageImp::setIsAllowed() const {
  LOG_DEBUG_F(GL, "PersistentStorageImp::setIsAllowed() isInWriteTran=%d, hasReplicaConfig=%d",
              isInWriteTran(), hasReplicaConfig());
  return (isInWriteTran() && hasReplicaConfig());
}

bool PersistentStorageImp::getIsAllowed() const {
  LOG_DEBUG_F(GL, "PersistentStorageImp::getIsAllowed() isInWriteTran=%d", isInWriteTran());
  return (!isInWriteTran());
}
```

There needs to be more work done to figure out what the true change is.
The recommended fix in #194 will not work for the reasons mentioned
above. A regression test that illustrates the problem is coming in a
follow up commit. We can use that to validate any real fixes.