Add memorydb transactions to storage library
Support for strict serializable transactions was added to the in memory
db. The transactions work by taking an exclusive lock over the in-memory
map holding the keys and values when the transaction is created. Writes
(Put and Remove) are put into a shadow map inside the transaction and
only applied on transaction commit. Reads inside the transaction first
look in the shadow map to see if the transaction modified the given
key and return the value in the shadow map if so. Otherwise they go to
the underlying map via the memorydb client.

Because the transactions take an exclusive lock for simplicity,
transactions should be short lived. We don't anticipate much concurrency
or complex transactions in the near future so this should be fine.

The storage tests were also modified to test the memorydb as well as
rocksdb. The multiIO tests verify the transactions work as expected.