Fix race when starting metrics server in skvbc
The metrics server was being started before the replica was created in
the SimpleKVBCTests TesterReplica. In release builds this could cause
requests to come in for the metrics server where it would return empty
json metrics because the replica hadn't yet registered it's metrics with the
aggregator provided by the replication server.

This change makes it so that the replica is constructed and metrics get
registered with the aggregator provided by the metrics server before it
gets started. This way no replies will be sent without metrics from the
metrics server.

The test was failing in release builds every other time when RocksDB
storage was enabled. It passed 19 times in a row after this change.