Restore default C++ compiler to clang++ when compiling.
Commit 53c6b149360f ("Enhance build via direct use of cmake (#56)") made
clang++ the default C++ compiler on Unix-like systems, but commit
b8bb68ac500c ("1. removed 3rd-party cmake FindLog4cplus for possible...")
reverted the setting.  I think that the latter was a mistake or an
oversight, because the commit message didn't mention the change and left
the comment that said that clang++ was the default.