Run pyclient_tests in verbose mode for CI
I'm having trouble reproducing failure of this test, but it fails
periodically in CI (See #212). Since it only takes about 2 seconds to run, run it
in verbose mode so we can see the stack trace. We may want to turn on
verbose mode for all tests in the future, but we need to ensure that
doesn't slow down runs significantly or exceed the 4MB log limit. We ran
into that problem prior.